+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
+## 2.6.2
+
+## Enhancements
+- Adds `Superwall.instance.consume(purchaseToken)` method to help easily consume in-app purchases
+
+## Fixes
+- Fixes issue with deeplink params not being handled properly in some cases
+- Fixes issue with Drawer and Modal displays on Android 14 Samsung devices
+- Fixes selection issue with some OTP, ensures after consuming the Status is synced
+
 ## 2.6.1
 
 ## Enhancements